+++ pashu-erp/packages/api/app/models/feed.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import String, Float, Boolean, DateTime, Enum, Numeric, text, func
+from sqlalchemy import String, Boolean, DateTime, Enum, Numeric, text, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -28,8 +28,8 @@ class FeedIngredient(Base):
     category: Mapped[str] = mapped_column(
         Enum(FeedCategory, name="feed_category"), nullable=False
     )
-    protein_pct: Mapped[float] = mapped_column(Float, nullable=False)
-    energy_kcal: Mapped[float] = mapped_column(Float, nullable=False)
+    protein_pct: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
+    energy_kcal: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
     cost_per_kg: Mapped[float] = mapped_column(Numeric(8, 2), nullable=False)
     availability_season: Mapped[str | None] = mapped_column(String(100), nullable=True)
     locally_available: Mapped[bool] = mapped_column(Boolean, default=True, server_default="true")
